Change pistons to not construct extra block entities
This fixes an issue with (at least) AE2 that's caused by creating a new
block entity instead of loading the data into the one created by vanilla
when the block state is set in the world.

This incidentally slightly changes the semantics of the IPistonCallback
onPistonMovementFinished method, causing it to be called on the block
entity that is in the world [after the data has been loaded into it]
instead of on a fresh block entity that will be replacing the one that's
in the world.